task 8

+++ 0x1E-search_algorithms/103-exponential.c
@@ -11,7 +11,7 @@
 
 int exponential_search(int *array, size_t size, int value)
 {
-	int bound = 1;
+	size_t bound = 1;
 
 	if (array == NULL)
 		return (-1);
